requirements.txt
streamlit==1.8.0
opencv-python-headless==4.5.3.56
numpy==1.21.1
Pillow==8.1.0
